+++ test/libsolidity/syntaxTests/tupleAssignments/fixed_byte_array_assignment_in_tuple.sol
@@ -1,0 +1,7 @@
+contract A {
+    function f() pure public {
+        bytes1 b;
+        bytes1 a;
+        (((, , b))) = (1, 2, a);
+    }
+}