+++ libsolidity/codegen/YulUtilFunctions.h
@@ -73,8 +73,9 @@ public:
 
 	/// @returns a function that copies raw bytes of dynamic length from calldata
 	/// or memory to memory.
-	/// Pads with zeros and might write more than exactly length.
-	std::string copyToMemoryFunction(bool _fromCalldata);
+	/// @params _cleanup If true, pads with zeros up to the 32 byte boundary after the specified length
+	/// signature: (src, dst, length) ->
+	std::string copyToMemoryFunction(bool _fromCalldata, bool _cleanup);
 
 	/// @returns the name of a function that copies a string literal to memory
 	/// and returns a pointer to the memory area containing the string literal.